Урок 6

+++ Lesson6Task2/Fruit.java
@@ -1,8 +1,8 @@
 package Lesson6Task2;
 
 public abstract class Fruit {
-    private Integer weight = 0;
-    public Integer getWeight() {
+    private double weight = 0d;
+    public double getWeight() {
         return weight;
     }
 }
